Added 'options' parameter to Squire; updated block tag references
All creating and comparing with 'DIV' now do so with options.blockTag (and
options.blockProperties where necessary), so that an API user can preperly
override the tag. The exceptions are insertHTML and onPaste. `instances`
array is attached to window so that it is accessible if the plugin is
wrapped in a function (module.exports, etc.); the links to Squire
instances are added to it immediately so that they are accessible ASAP in
functions line fixCursor