Avoid mangling the `-- ` signature delimiter
As Squire is used for email clients a lot (I'm a happy Fastmail
customer), and those might insert the signature into new emails
with `setHTML`, it would be good if it would leave `-- `
[signature delimiter](https://en.wikipedia.org/wiki/Signature_block#Standard_delimiter)
intact.

You could argue it's not very nice to 'special-case' the delimiter like
this, though on the other hand it really *is* a special case and all
other trailing whitespace does typically deserve to be truncated.