Treat ZWS as white space
We use the ZWS character as a placeholder to make elements focusable,
but then we strip them again afterwards. So if we come across one, we
should just ignore it, treating it like insignificant white space.

This fixes an issue where if you pasted a link into an empty editor,
then hit enter, it wouldn't break the line, as it would split the link
at the end, creating an empty <a>, which gets a ZWS inserted in it,
which caused it to not add a <br> to the new line when it needed one.