Fix drag-drop text can delete sections
As reported by @lksoft on Github:

When a section of text is selected inside of a text node and then is
dragged to another location in the editor, the text from the end of the
original selection to the end of the paragraph is deleted when the move
is performed. The reason is because when the undo state is being changed
the resulting range (after the _getRangeAndRemoveBookmark()) is not used
to reset the selection, which ends up selecting the rest of the
paragraph and thus the move removes the whole selection.