Don't move range when copying data
I don't think this should make any difference when we're just cloning
the contents — it should only change how many nodes are split, but
that would only apply if deleting the contents, which we're not.

Meanwhile, this works around a bug in Firefox where for some reason it
sometimes fails to clone all the contents of the range after you move
the boundaries (it misses the bit in the middle of the selection). I
haven't dived into the exact reproduction circumstances, as I don't have
time right now and this workaround seems fine).